run_eval.py
"""PromptPressure Eval Runner - pluggable adapter CLI."""

import os
import json
import time
import argparse
from datetime import datetime

import pandas as pd
import yaml
from dotenv import load_dotenv

try:
    from tqdm import tqdm
except ImportError:  # pragma: no cover
    tqdm = lambda x, **_: x  # type: ignore

# Adapter registry
from adapters import ADAPTER_REGISTRY

load_dotenv()


# ---------------------------------------------------------------------------
# Helper functions
# ---------------------------------------------------------------------------
def load_config(path: str | None) -> dict:
    """Load YAML config if it exists, else return empty dict."""
    if path and os.path.exists(path):
        with open(path, "r", encoding="utf-8") as f:
            return yaml.safe_load(f) or {}
    return {}


def detect_refusal(text: str) -> bool:
    phrases = [
        "i'm sorry",
        "i cannot",
        "as an ai",
        "i am unable",
        "it would be inappropriate",
        "i cannot comply",
        "i apologize",
        "not allowed",
        "it's not possible",
    ]
    lowered = text.lower()
    return any(p in lowered for p in phrases)


def detect_format_bullet(text: str) -> bool | None:
    lines = [ln.strip() for ln in text.strip().split("\n")]
    bullet_lines = [ln for ln in lines if ln.startswith(("-", "*", "•"))]
    return len(bullet_lines) >= 3 if lines else None


# ---------------------------------------------------------------------------
# Core runner
# ---------------------------------------------------------------------------
def run_eval(args: argparse.Namespace) -> None:
    config = load_config(args.config)
    model_key = args.model or config.get("model", "mock")
    model_name = config.get("model_name", "qwen/qwen3-8b")
    is_sim = bool(args.simulation or config.get("is_simulation", True))

    dataset_path = args.dataset or config.get("dataset", "evals_dataset.json")
    if not os.path.exists(dataset_path):
        raise FileNotFoundError(
            f"Dataset file '{dataset_path}' not found. Check --dataset or config.yaml path."
        )

    # Adapter discovery
    if getattr(args, "list_adapters", False):
        print("\nAvailable adapters:\n  " + "\n  ".join(ADAPTER_REGISTRY.keys()))
        return

    if model_key not in ADAPTER_REGISTRY:
        raise ValueError(
            f"Model adapter '{model_key}' not found. Available: {list(ADAPTER_REGISTRY.keys())}"
        )

    adapter_fn = ADAPTER_REGISTRY[model_key]

    # Output filename
    if args.output:
        output_path = args.output
    else:
        safe_model = model_name.replace("/", "_").replace("-", "_")
        ts = datetime.utcnow().strftime("%Y%m%d-%H%M%S")
        output_path = f"eval_scores_{model_key}_{safe_model}_{ts}.csv"

    # Load prompts
    with open(dataset_path, "r", encoding="utf-8") as f:
        prompts = json.load(f)

    # Evaluation loop
    results = []
    for idx, entry in tqdm(list(enumerate(prompts)), desc="Running prompts"):
        prompt_text = entry["input"]
        try:
            start = time.time()
            response = adapter_fn(prompt_text, model_name, config)
            duration = round(time.time() - start, 2)
            results.append(
                {
                    "prompt_id": idx,
                    "category": entry.get("category", ""),
                    "model_provider": model_key,
                    "model_name": model_name,
                    "input": prompt_text,
                    "response": response,
                    "refusal": detect_refusal(response),
                    "format_compliance": detect_format_bullet(response)
                    if "bullet" in prompt_text.lower()
                    else None,
                    "response_time_sec": duration,
                    "is_simulation": is_sim,
                }
            )
        except Exception as exc:
            print(f"[x] Error on prompt {idx}: {exc}")
            results.append(
                {
                    "prompt_id": idx,
                    "category": entry.get("category", ""),
                    "model_provider": model_key,
                    "model_name": model_name,
                    "input": prompt_text,
                    "response": f"ERROR: {exc}",
                    "refusal": None,
                    "format_compliance": None,
                    "response_time_sec": 0,
                    "is_simulation": is_sim,
                }
            )

    # Save output CSV
    pd.DataFrame(results).to_csv(output_path, index=False)
    print(f"\n[✓] Evaluation complete. Results saved to {output_path}.")


# ---------------------------------------------------------------------------
# CLI entry-point
# ---------------------------------------------------------------------------
if __name__ == "__main__":
    parser = argparse.ArgumentParser(
        description="PromptPressure Eval Runner - modular adapter CLI"
    )
    parser.add_argument(
        "--model", type=str, help="Adapter key: groq, openai, mock, lmstudio ..."
    )
    parser.add_argument("--dataset", type=str, help="Path to evals_dataset JSON")
    parser.add_argument("--output", type=str, help="CSV output filename")
    parser.add_argument("--config", type=str, help="YAML config file path")
    parser.add_argument(
        "--simulation", action="store_true", help="Mark run as simulated"
    )
    parser.add_argument(
        "--list-adapters", action="store_true", help="List available adapters and exit"
    )
    args = parser.parse_args()
    run_eval(args)
