requirements.txt
pandas
requests
pyyaml
python-dotenv
tqdm